update v1.0.3 加入了更改弹幕中心点，中心扩缩，旋转角度，冷却时间精确到0.01秒…

+++ Command.cs
@@ -21,7 +21,7 @@ public class Command
         {
             if (args.Parameters.Count == 2)
             {
-                if (double.TryParse(args.Parameters[0], out var damage) && int.TryParse(args.Parameters[1], out var cooldown))
+                if (double.TryParse(args.Parameters[0], out var damage) && double.TryParse(args.Parameters[1], out var cooldown))
                 {
                     Config.ExtraDamage = damage;
                     Config.Cooldown = cooldown;
